move signal.go to common
