+++ src/cmd/preview.go
@@ -34,7 +34,7 @@ If the optional argument SAMPLES_COUNT=0 this will print out everything.`,
 		cobra.CheckErr(err)
 
 		queue := make(chan opslevel_jq_parser.ServiceRegistration, 1)
-		ctx := InitSignalHandler(context.Background(), queue)
+		ctx := common.InitSignalHandler(context.Background(), queue)
 		client := createOpslevelClient()
 		common.SyncCache(client)
 		common.SetupControllers(config, queue, 0, ctx)
