+++ src/cmd/reconcile.go
@@ -25,7 +25,7 @@ var reconcileCmd = &cobra.Command{
 		cobra.CheckErr(err)
 
 		queue := make(chan opslevel_jq_parser.ServiceRegistration, 1)
-		ctx := InitSignalHandler(context.Background(), queue)
+		ctx := common.InitSignalHandler(context.Background(), queue)
 		client := createOpslevelClient()
 		common.SyncCache(client)
 		common.SyncCaches(createOpslevelClient(), resync)
