+++ src/common/signal.go
@@ -1,4 +1,4 @@
-package cmd
+package common
 
 import (
 	"context"
